Classi BL e tipi
-Per i cast, usati i metodi Convert To per maggior robustezza del codice
-Cambiata ottava su ClsNotaMusicale da int a byte
-GetAll su ClsNotaMusicaleBL e test del metodo su FrmHome

+++ NegozioStrumentiMusicali_Cappelloni-DiBernardo/BL/ClsPiattoBL.cs
@@ -181,7 +181,7 @@ namespace NegozioStrumentiMusicali
                     while(_dataReader.Read()) //Se ce li ha li leggo tutti
                     {
                         ClsPiatto _piatto = new ClsPiatto();
-                        _piatto.ID = Convert.ToUInt32(_dataReader["ID"]);
+                        _piatto.ID = Convert.ToInt64(_dataReader["ID"]);
                         _piatto.Materiale = (ClsPiatto.eMATERIALE)Enum.Parse
                         (
                             typeof(ClsPiatto.eMATERIALE),
